add test in a new test page to run it manually
